started a larger refactor to clarify a few major codepaths that are getting comple
