+++ client/src/components/room/room.tsx
@@ -20,8 +20,9 @@ import { useUserStore } from "../../stores/userStore";
 import AdminPanel from "../account/adminPanel";
 import ArtistRoom from "../artistRooms/artistRoom";
 import Layout from "./layout";
-import RoomNameGate, { RoomStatusGate } from "./roomGate";
-import StreamGate from "./streamGate";
+import RoomStatusGate  from "../gates/roomStatusGate";
+import RoomExistsGate  from "../gates/roomExistsGate";
+import StreamGate from "../gates/streamGate";
 
 const Room: React.FC<{ roomID: string; season?: number }> = ({
   roomID,
@@ -92,7 +93,7 @@ const Room: React.FC<{ roomID: string; season?: number }> = ({
 
   return (
     <Layout roomColor={roomColor}>
-      <RoomNameGate id={roomID as string}>
+      <RoomExistsGate id={roomID as string}>
         <StreamGate
           roomID={roomID}
           streamPlaybackID={roomPlaybackId}
@@ -107,7 +108,7 @@ const Room: React.FC<{ roomID: string; season?: number }> = ({
             </>
           )}
         </StreamGate>
-      </RoomNameGate>
+      </RoomExistsGate>
     </Layout>
   );
 };
